feat: add user validation and authentication/permissions

+++ src/projects/projects.service.ts
@@ -11,23 +11,34 @@ export class ProjectsService {
     private projectsRepository: Repository<Project>,
   ) { }
 
-  create(project: ProjectDTO): Promise<Project> {
-    return this.projectsRepository.save(project)
+  create(userId: number, project: ProjectDTO): Promise<Project> {
+    return this.projectsRepository.save({
+      name: project.name,
+      userId: userId
+    })
   }
 
-  findAll(): Promise<Project[]> {
-    return this.projectsRepository.find();
+  findAllByUser(user: number): Promise<Project[]> {
+    return this.projectsRepository.find({ where: { userId: user }, relations: ['user'] });
   }
 
-  findOne(id: string): Promise<Project> {
-    return this.projectsRepository.findOne(id)
+  async findOne(user: number, id: string): Promise<Project> {
+    const project: Project = await this.projectsRepository.findOne(id)
+
+    if (project.userId !== user) throw new Error('User not allowed to access this project')
+
+    return project
   }
 
   update(id: string, updateProjectDTO: ProjectDTO) {
     return this.projectsRepository.update(id, updateProjectDTO)
   }
 
-  async remove(id: string): Promise<void> {
+  async remove(user: number, id: string): Promise<void> {
+    const project: Project = await this.projectsRepository.findOne(id)
+
+    if (project.userId !== user) throw new Error('User not allowed to delete this project')
+
     await this.projectsRepository.delete(id);
   }
 }
